+++ src/main/java/AWSAccessors/JavaLocalGrailsUnit.java
@@ -1,5 +1,7 @@
 package AWSAccessors;
 
+import java.math.BigDecimal;
+
 public class JavaLocalGrailsUnit {
     public long id;
     public String name;
@@ -8,7 +10,7 @@ public class JavaLocalGrailsUnit {
     public double height;
     public String type;
     public int floor;
-    public double price;
+    public BigDecimal price;
     public long facilityId;
 
     public String rateType;
@@ -18,7 +20,7 @@ public class JavaLocalGrailsUnit {
 
     }
 
-    public JavaLocalGrailsUnit(long id, String name, double width, double depth, double height, String type, int floor, double price, long facilityId, String rateType)
+    public JavaLocalGrailsUnit(long id, String name, double width, double depth, double height, String type, int floor, BigDecimal price, long facilityId, String rateType)
     {
         this.id = id;
         this.name = name;
